requirements.txt
psutil
py-cpuinfo
pandas
requests
